+++ tests/TypeConstraintTest.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace League\JsonGuard\Test {
+    use League\JsonGuard\Constraints\Type;
+
+    class TypeConstraintTest extends \PHPUnit_Framework_TestCase
+    {
+        /**
+         * @runInSeparateProcess
+         */
+        public function testTypeStringPassesWhenValidAndBcMathIsNotInstalled()
+        {
+            require_once __DIR__ . '/stubs/constraint_function_exists.php';
+            $this->assertNull(Type::validate('hello world', 'string'));
+        }
+    }
+}